fix typos for moving stuff around
